+++ CHANGELOG.md
@@ -104,6 +104,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
       it is actually a directory (same as the source path). Otherwise, if the
       *target* path has a trailing slash then the *source* path must be a
       directory (otherwise you get `ENOTDIR`).
+- opath resolver: we now return `ELOOP` when we run into a symlink that came
+  from mount with the `MS_NOSYMFOLLOW` set, to match the behaviour of `openat2`.
 
 ### Changed ###
 - syscalls: switch to rustix for most of our syscall wrappers to simplify how